Orthogonal weight update (GPU only)
Regularized/Normalized diffs computed by solver will be projected on to
the Tangent Space in the Stiefel manifold of the orthogonal weights in
RecursiveConvLayer, and new weights will be computed using Cayley's
transform. This will ensure that the weights always remain orthogonal
in a natural way, whilst simultaneously optimizing the problem at hand.
Currently implemented only for GPU using cusolverDn.
(Tests that are no longer required are removed.)